move common fields and methods from Suite/Test to Item
